+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change link_announcer.py to only warn on connection errors
 - Change user lookup logic in last.fm plugin
 - Refactor minecraft_ping plugin for updated mcstatus library
+- Expand youtube.py error information
 ### Fixed
 - Fix matching exception in horoscope test
 - Fix youtube.py ISO time parse